Split into multiple TypeScript projects
