+++ src/lib/code-sample-editor-layout.ts
@@ -6,7 +6,7 @@ import {
   property,
   query,
 } from 'lit-element';
-import {TemplateResult} from 'lit-html';
+import { TemplateResult } from 'lit-html';
 
 @customElement('code-sample-editor-layout')
 export class CodeSampleEditorLayout extends LitElement {
@@ -51,13 +51,13 @@ export class CodeSampleEditorLayout extends LitElement {
     `;
   }
 
-  @property({type: Boolean})
+  @property({ type: Boolean })
   addFilePrompt = false;
 
   @query('#filenameInput')
   filenameInput?: HTMLInputElement;
 
-  private onSubmit(e: Event) {
+  private onSubmit(_e: Event) {
     this.dispatchEvent(
       new CustomEvent('save', {
         bubbles: true,
@@ -66,7 +66,7 @@ export class CodeSampleEditorLayout extends LitElement {
     );
   }
 
-  private onCreateFileClick(e: Event) {
+  private onCreateFileClick(_e: Event) {
     this.dispatchEvent(
       new CustomEvent('create-file', {
         bubbles: true,
@@ -78,7 +78,7 @@ export class CodeSampleEditorLayout extends LitElement {
     this.addFilePrompt = false;
   }
 
-  private onAddFileClick(e: Event) {
+  private onAddFileClick(_e: Event) {
     this.addFilePrompt = true;
   }
